backend/requirements.txt
strands-agents
fastapi
uvicorn
python-dotenv
ddtrace
boto3
requests
anthropic
pydantic
neo4j
